+++ arch/powerpc/sysdev/Kconfig
@@ -2,6 +2,12 @@
 # see Documentation/kbuild/kconfig-language.txt.
 #
 
+config LSI_PCI
+	bool "LSI ACP34XX PCIe support"
+	default y
+        help
+          PCIe driver support for LSI's ACP34xx
+
 config PPC4xx_PCI_EXPRESS
 	bool
 	depends on PCI && 4xx
